fix(core): use a queue to process events in event routes
